[refactor]: Move metrics related functionality into MetricsReporter
