+++ cli/src/lib.rs
@@ -11,6 +11,8 @@ use std::{path::PathBuf, sync::Arc};
 use clap::Parser;
 use color_eyre::eyre::{eyre, Result, WrapErr};
 use iroha_config::parameters::{actual::Root as Config, user::CliContext};
+#[cfg(feature = "telemetry")]
+use iroha_core::metrics::MetricsReporter;
 use iroha_core::{
     block_sync::{BlockSynchronizer, BlockSynchronizerHandle},
     gossiper::{TransactionGossiper, TransactionGossiperHandle},
@@ -214,6 +216,7 @@ impl Iroha {
         );
 
         let kura = Kura::new(&config.kura)?;
+        let kura_thread_handler = Kura::start(Arc::clone(&kura));
         let live_query_store_handle = LiveQueryStore::from_config(config.live_query_store).start();
 
         let block_count = kura.init()?;
@@ -254,7 +257,13 @@ impl Iroha {
         #[cfg(feature = "telemetry")]
         Self::start_telemetry(&logger, &config).await?;
 
-        let kura_thread_handler = Kura::start(Arc::clone(&kura));
+        #[cfg(feature = "telemetry")]
+        let metrics_reporter = MetricsReporter::new(
+            Arc::clone(&state),
+            network.clone(),
+            kura.clone(),
+            queue.clone(),
+        );
 
         let start_args = SumeragiStartArgs {
             sumeragi_config: config.sumeragi.clone(),
@@ -269,6 +278,7 @@ impl Iroha {
                 public_key: config.genesis.public_key().clone(),
             },
             block_count,
+            dropped_messages: metrics_reporter.metrics().dropped_messages.clone(),
         };
         // Starting Sumeragi requires no async context enabled
         let sumeragi = tokio::task::spawn_blocking(move || SumeragiHandle::start(start_args))
@@ -322,11 +332,11 @@ impl Iroha {
             Arc::clone(&queue),
             events_sender,
             Arc::clone(&notify_shutdown),
-            #[cfg(feature = "telemetry")]
-            sumeragi.clone(),
             live_query_store_handle,
             Arc::clone(&kura),
             Arc::clone(&state),
+            #[cfg(feature = "telemetry")]
+            metrics_reporter,
         );
 
         Self::spawn_config_updates_broadcasting(kiso.clone(), logger.clone());
