+++ core/src/lib.rs
@@ -6,6 +6,7 @@ pub mod executor;
 pub mod gossiper;
 pub mod kiso;
 pub mod kura;
+pub mod metrics;
 pub mod query;
 pub mod queue;
 pub mod smartcontracts;
