+++ core/src/sumeragi/mod.rs
@@ -4,18 +4,16 @@
 use std::{
     fmt::{self, Debug, Formatter},
     sync::{mpsc, Arc},
-    time::{Duration, Instant, SystemTime},
+    time::{Duration, Instant},
 };
 
-use eyre::{Result, WrapErr as _};
+use eyre::Result;
 use iroha_config::parameters::actual::{Common as CommonConfig, Sumeragi as SumeragiConfig};
 use iroha_crypto::{KeyPair, SignatureOf};
 use iroha_data_model::{block::SignedBlock, prelude::*};
 use iroha_genesis::GenesisNetwork;
 use iroha_logger::prelude::*;
-use iroha_telemetry::metrics::Metrics;
 use network_topology::{Role, Topology};
-use storage::storage::StorageReadOnly;
 
 use crate::{
     block::ValidBlock,
@@ -29,30 +27,14 @@ pub mod message;
 pub mod network_topology;
 pub mod view_change;
 
-use parking_lot::Mutex;
-
 use self::{message::*, view_change::ProofChain};
 use crate::{kura::Kura, prelude::*, queue::Queue, EventsSender, IrohaNetwork, NetworkMessage};
-
-/*
-The values in the following struct are not atomics because the code that
-operates on them assumes their values does not change during the course of
-the function.
-*/
-#[derive(Debug)]
-struct LastUpdateMetricsData {
-    block_height: u64,
-}
 
 /// Handle to `Sumeragi` actor
 #[derive(Clone)]
 pub struct SumeragiHandle {
-    state: Arc<State>,
-    metrics: Metrics,
-    last_update_metrics_mutex: Arc<Mutex<LastUpdateMetricsData>>,
-    network: IrohaNetwork,
-    kura: Arc<Kura>,
-    queue: Arc<Queue>,
+    /// Counter for amount of dropped messages by sumeragi
+    dropped_messages: iroha_telemetry::metrics::IntCounter,
     _thread_handle: Arc<ThreadHandler>,
     // Should be dropped after `_thread_handle` to prevent sumeargi thread from panicking
     control_message_sender: mpsc::SyncSender<ControlFlowMessage>,
@@ -60,119 +42,10 @@ pub struct SumeragiHandle {
 }
 
 impl SumeragiHandle {
-    /// Update the metrics on the world state view.
-    ///
-    /// # Errors
-    /// - Domains fail to compose
-    ///
-    /// # Panics
-    /// - If either mutex is poisoned
-    #[allow(clippy::cast_precision_loss)]
-    pub fn update_metrics(&self) -> Result<()> {
-        let online_peers_count: u64 = self
-            .network
-            .online_peers(
-                #[allow(clippy::disallowed_types)]
-                std::collections::HashSet::len,
-            )
-            .try_into()
-            .expect("casting usize to u64");
-
-        let state_view = self.state.view();
-
-        let mut last_guard = self.last_update_metrics_mutex.lock();
-
-        let start_index = last_guard.block_height;
-        {
-            let mut block_index = start_index;
-            while block_index < state_view.height() {
-                let Some(block) = self.kura.get_block_by_height(block_index + 1) else {
-                    break;
-                };
-                block_index += 1;
-                let mut block_txs_accepted = 0;
-                let mut block_txs_rejected = 0;
-                for tx in block.transactions() {
-                    if tx.error.is_none() {
-                        block_txs_accepted += 1;
-                    } else {
-                        block_txs_rejected += 1;
-                    }
-                }
-
-                self.metrics
-                    .txs
-                    .with_label_values(&["accepted"])
-                    .inc_by(block_txs_accepted);
-                self.metrics
-                    .txs
-                    .with_label_values(&["rejected"])
-                    .inc_by(block_txs_rejected);
-                self.metrics
-                    .txs
-                    .with_label_values(&["total"])
-                    .inc_by(block_txs_accepted + block_txs_rejected);
-                self.metrics.block_height.inc();
-            }
-            last_guard.block_height = block_index;
-        }
-
-        let new_tx_amounts = {
-            let mut new_buf = Vec::new();
-            core::mem::swap(&mut new_buf, &mut state_view.new_tx_amounts.lock());
-            new_buf
-        };
-
-        for amount in &new_tx_amounts {
-            self.metrics.tx_amounts.observe(*amount);
-        }
-
-        #[allow(clippy::cast_possible_truncation)]
-        if let Some(timestamp) = state_view.genesis_timestamp() {
-            let curr_time = SystemTime::now()
-                .duration_since(SystemTime::UNIX_EPOCH)
-                .expect("Failed to get the current system time");
-
-            // this will overflow in 584942417years.
-            self.metrics.uptime_since_genesis_ms.set(
-                (curr_time - timestamp)
-                    .as_millis()
-                    .try_into()
-                    .expect("Timestamp should fit into u64"),
-            )
-        };
-
-        self.metrics.connected_peers.set(online_peers_count);
-
-        self.metrics
-            .domains
-            .set(state_view.world().domains().len() as u64);
-        for domain in state_view.world().domains_iter() {
-            self.metrics
-                .accounts
-                .get_metric_with_label_values(&[domain.id().name.as_ref()])
-                .wrap_err("Failed to compose domains")?
-                .set(domain.accounts.len() as u64);
-        }
-
-        self.metrics
-            .view_changes
-            .set(state_view.latest_block_view_change_index());
-
-        self.metrics.queue_size.set(self.queue.tx_len() as u64);
-
-        Ok(())
-    }
-
-    /// Access node metrics.
-    pub fn metrics(&self) -> &Metrics {
-        &self.metrics
-    }
-
     /// Deposit a sumeragi control flow network message.
     pub fn incoming_control_flow_message(&self, msg: ControlFlowMessage) {
         if let Err(error) = self.control_message_sender.try_send(msg) {
-            self.metrics.dropped_messages.inc();
+            self.dropped_messages.inc();
             error!(
                 ?error,
                 "This peer is faulty. \
@@ -184,7 +57,7 @@ impl SumeragiHandle {
     /// Deposit a sumeragi network message.
     pub fn incoming_block_message(&self, msg: BlockMessage) {
         if let Err(error) = self.message_sender.try_send(msg) {
-            self.metrics.dropped_messages.inc();
+            self.dropped_messages.inc();
             error!(
                 ?error,
                 "This peer is faulty. \
@@ -255,6 +128,7 @@ impl SumeragiHandle {
             network,
             genesis_network,
             block_count: BlockCount(block_count),
+            dropped_messages,
         }: SumeragiStartArgs,
     ) -> SumeragiHandle {
         let (control_message_sender, control_message_receiver) = mpsc::sync_channel(100);
@@ -353,16 +227,9 @@ impl SumeragiHandle {
 
         let thread_handle = ThreadHandler::new(Box::new(shutdown), thread_handle);
         SumeragiHandle {
-            state,
-            network,
-            queue,
-            kura,
+            dropped_messages,
             control_message_sender,
             message_sender,
-            metrics: Metrics::default(),
-            last_update_metrics_mutex: Arc::new(Mutex::new(LastUpdateMetricsData {
-                block_height: 0,
-            })),
             _thread_handle: Arc::new(thread_handle),
         }
     }
@@ -428,6 +295,7 @@ pub struct SumeragiStartArgs {
     pub network: IrohaNetwork,
     pub genesis_network: GenesisWithPubKey,
     pub block_count: BlockCount,
+    pub dropped_messages: iroha_telemetry::metrics::IntCounter,
 }
 
 /// Optional genesis paired with genesis public key for verification
